Appeasement cast for 32-bit maxhines.

+++ ntpshmmon.c
@@ -89,9 +89,9 @@ int main(int argc, char **argv)
 		if (shm_stat.tvc.tv_sec != tick[i].tv_sec || shm_stat.tvc.tv_nsec != tick[i].tv_nsec) {
 		    printf("sample %s %ld.%09ld %ld.%09ld %ld.%09ld %d %3d\n",
 			   ntp_name(i),
-			   shm_stat.tvc.tv_sec, shm_stat.tvc.tv_nsec,
-			   shm_stat.tvr.tv_sec, shm_stat.tvr.tv_nsec,
-			   shm_stat.tvt.tv_sec, shm_stat.tvt.tv_nsec,
+			   (long)shm_stat.tvc.tv_sec, shm_stat.tvc.tv_nsec,
+			   (long)shm_stat.tvr.tv_sec, shm_stat.tvr.tv_nsec,
+			   (long)shm_stat.tvt.tv_sec, shm_stat.tvt.tv_nsec,
 			   shm_stat.leap, shm_stat.precision);
 		    tick[i] = shm_stat.tvc;
 		    --nsamples;
